Moving WhatAmI into config
