+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/config/WhatAmI.kt
@@ -12,7 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
-package io.zenoh.scouting
+package io.zenoh.config
 
 /**
  * WhatAmI
