+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/jni/JNIScout.kt
@@ -21,7 +21,7 @@ import io.zenoh.jni.callbacks.JNIScoutCallback
 import io.zenoh.config.ZenohId
 import io.zenoh.scouting.Hello
 import io.zenoh.scouting.Scout
-import io.zenoh.scouting.WhatAmI
+import io.zenoh.config.WhatAmI
 
 /**
  * Adapter class to handle the interactions with Zenoh through JNI for a [io.zenoh.scouting.Scout]
